+++ packages/common/sentry/package.json
@@ -8,7 +8,7 @@
   "author": "DXOS.org",
   "main": "dist/lib/node/index.cjs",
   "browser": {
-    "./src/node-util.ts": "./src/browser-util.ts",
+    "./src/node.ts": "./src/browser.ts",
     "./testing/testkit/index.ts": "./testing/testkit/browser.ts",
     "./dist/lib/node/index.cjs": "./dist/lib/browser/index.mjs"
   },
