+++ packages/common/sentry/src/index.ts
@@ -2,7 +2,8 @@
 // Copyright 2022 DXOS.org
 //
 
-export * from './init';
+export * from './node';
+export * from './user';
 export * from './types';
 
 // TODO(mykola): Tracing does not work in the node. Delete this export.
