--- packages/common/sentry/src/init.ts
@@ -1,47 +1,0 @@
-//
-// Copyright 2022 DXOS.org
-//
-
-import { log } from '@dxos/log';
-
-import { naturalAddBreadcrumb, naturalCaptureException } from './node-util';
-
-export * from './node-util';
-// Adds user sentry processor to @dxos/log framework.
-export * from './user';
-
-/**
- * Records a new breadcrumb which will be attached to future events.
- *
- * Breadcrumbs will be added to subsequent events to provide more context on user's actions prior to an error or crash.
- *
- * https://docs.sentry.io/platforms/javascript/enriching-events/breadcrumbs/
- *
- * @param breadcrumb — The breadcrumb to record.
- */
-export const addBreadcrumb: typeof naturalAddBreadcrumb = (breadcrumb) => {
-  try {
-    naturalAddBreadcrumb(breadcrumb);
-    log('add breadcrumb', breadcrumb);
-  } catch (err) {
-    log.catch('Failed to add breadcrumb', err);
-  }
-};
-
-/**
- * Captures an exception event and sends it to Sentry.
- *
- * @param exception — An exception-like object.
- * @param captureContext — Additional scope data to apply to exception event.
- * @returns — The generated eventId.
- */
-export const captureException: typeof naturalCaptureException = (exception, captureContext) => {
-  try {
-    const eventId = naturalCaptureException(exception, captureContext);
-    log('capture exception', { exception, eventId, ...captureContext });
-    return eventId;
-  } catch (err) {
-    log.catch('Failed to capture exception', err);
-    return 'unknown';
-  }
-};
