+++ packages/common/sentry/src/node.ts
@@ -3,7 +3,12 @@
 //
 
 import { CaptureConsole } from '@sentry/integrations';
-import { init as naturalInit, setTag } from '@sentry/node';
+import {
+  init as naturalInit,
+  setTag,
+  addBreadcrumb as naturalAddBreadcrumb,
+  captureException as naturalCaptureException,
+} from '@sentry/node';
 import type { Event } from '@sentry/node';
 
 import { log } from '@dxos/log';
@@ -11,13 +16,7 @@ import { log } from '@dxos/log';
 import { InitOptions } from './types';
 
 // Polyfill export.
-export {
-  setTag,
-  setUser,
-  getCurrentHub,
-  addBreadcrumb as naturalAddBreadcrumb,
-  captureException as naturalCaptureException,
-} from '@sentry/node';
+export { setTag, setUser } from '@sentry/node';
 
 /**
  * To use this SDK, call the init function as early as possible in the main entry module.
@@ -63,4 +62,40 @@ const scrub = (event: Event) => {
       frame.filename = filename && filename[filename.length - 1];
     });
   });
+};
+
+/**
+ * Records a new breadcrumb which will be attached to future events.
+ *
+ * Breadcrumbs will be added to subsequent events to provide more context on user's actions prior to an error or crash.
+ *
+ * https://docs.sentry.io/platforms/javascript/enriching-events/breadcrumbs/
+ *
+ * @param breadcrumb — The breadcrumb to record.
+ */
+export const addBreadcrumb: typeof naturalAddBreadcrumb = (breadcrumb) => {
+  try {
+    naturalAddBreadcrumb(breadcrumb);
+    log('add breadcrumb', breadcrumb);
+  } catch (err) {
+    log.catch('Failed to add breadcrumb', err);
+  }
+};
+
+/**
+ * Captures an exception event and sends it to Sentry.
+ *
+ * @param exception — An exception-like object.
+ * @param captureContext — Additional scope data to apply to exception event.
+ * @returns — The generated eventId.
+ */
+export const captureException: typeof naturalCaptureException = (exception, captureContext) => {
+  try {
+    const eventId = naturalCaptureException(exception, captureContext);
+    log('capture exception', { exception, eventId, ...captureContext });
+    return eventId;
+  } catch (err) {
+    log.catch('Failed to capture exception', err);
+    return 'unknown';
+  }
 };
