+++ packages/common/sentry/src/sentry.test.ts
@@ -7,7 +7,7 @@ import waitForExpect from 'wait-for-expect';
 
 import { beforeAll, beforeEach, describe, test } from '@dxos/test';
 
-import * as Sentry from './init';
+import * as Sentry from './node';
 import { sentryTestkit } from '../testing';
 
 const { testkit, sentryTransport } = sentryTestkit();
